Add basic superevent functionality
After this triggers will be processed serially for creation
and modification of superevents. There is a dedicated gwcelery
queue for this purpose. Current implementation simply decides
preferred events based on FAR values and superevent time window
is set to ad hoc value read in from celery config. This will
worked on in later merge requests.