Implement use of glue.segments tools in dealing with events and superevents
The condition of an event lying inside a superevent window is now handled using
glue.segment tools. When the superevent list is pulled down from GraceDb, it is
changed to a glue.segments.segmentlist for easy operations like interstion with
events. Further operations like shift and protract on events could be easily
implemented using these tools.