Fix misuse of GitLab CI after_script section
I had been misusing the `after_script` section to run the
`docker build` commands. As a result, failures of `docker build`
were being mis-reported as successful CI jobs.

According to the GitLab CI help, the `after_script` section is
run regardless of whether the `script` section succeeds or fails,
and presumably if the `after_script` section fails, it has no
impact on whether the job as a whole is reported as a failure.

The `after_script` section is apparently not supposed to be used
for critical functionality of the job, but for non-critical
cleanup code.