Move orchestrator timeout inside the canvas
There is a controlled pause between the creation of a new
superevent and the generation of the preliminary alert so that the
choice of preferred event can stabilize. However, some tasks (such
as state vector checks) are not going to depend on the choice of
preferred event.

It turns out that we can set execution attributes like `countdown`
on a per-signature basis. By moving the pause inside the canvas,
we make it easier to add other tasks that can execute while the
countdown is proceeding.