Restrict state vector check to detectors that were online
Since we have to wait until the preferred event is known, the
state vector check had to move later in the canvas. And besides,
when we originally had the state vector check before
`get_preferred_event`, it caused the subsequent annotations to
never run. I'm not sure why; perhaps a bug in Celery.