Fix dropped VOEvents when sending keepalives
The GCN broker peeks at the first VOEvent in the redis queue,
and if there is none, sends a keepalive packet.

The GCN broker was unconditionally removing (popping) the first
VOEvent, even if it had sent a keepalive packet, rather than an
actual VOEvent. This was probably normally harmless, but created
the possibility of dropping a queued VOEvent if one was added
while sending a keepalive message.