Add condor.submit to run a job from a .sub file
Add the task `gwcelery.tasks.condor.submit` to run a condor job
from a .sub file.

By contrast, the older task which has been renamed to
`gwcelery.tasks.condor.check_output` runs a condor job
by building a .sub file on the fly in order to run the
command that is provided by the user.

This is in support of !228.