Disable GCN for all configurations except for production
Now that we want to have multiple deployments of GWCelery running
side by side on the same machine for development and production,
we need to disable the GCN broker and client for all but one
configuration. There can be only one GCN broker because it has to
bind to a port. There can be only one GCN client because of a
limitation in how GCN does IP address authentication.

Fixes #68.