Set redis socket in conftest
The nagios unit tests require starting a temporary Redis database
listening on a UNIX domain socket. The problem with setting the
Celery app's broker URL in the Nagios unit tests is that Celery
4.3.0 does lazy initialization of the broker connection, and it's
too late to change the broker URL by the time this test runs.

It is more reliable to set the broker URL just once, at the pytest
session's scope.