Expose events to the public prior to sending any kind of alert
Expose events to the public prior to sending any kind of alert:
preliminary, initial, update, or retraction. Previously this
behavior only occurred for preliminary alerts, which created the
unusual and undesirable possibility of a public GCN for an event
that is not public. As before, events are only exposed to the
public in the production environment, and not in the playground
environment.