Fix an incorrect assumption in get_instruments_in_ranking_statistic()
The function currently assumes that only detectors with a valid
chi^2 field contribute to the ranking of an event. This is not
true for PyCBC Live in O3, which can have empty chi^2 values even
though *all* detectors always contribute to the ranking.

This patch fixes the get_instruments_in_ranking_statistic()
function to always return all detectors for events from
PyCBC Live.