Fix copy-paste error: master_doc -> main_doc
The `master_doc` variable in `doc/conf.py` was accidentally
renamed to `main_doc` in #356. It is a Sphinx configuration
variable and is not related to the name of the main branch.

Instead of reverting it back to `master_doc`, rename it to
`root_doc`, which is the new name of this variable in Sphinx 4.
Consequently, we also now depend on Sphinx >= 4.